Provider: Setup and added appropriate ViewModel
