+++ lib/ui/shared/app_colors.dart
@@ -9,4 +9,19 @@ class MinitelColors {
   static const Color TerminalFgColor = Colors.white;
 
   static const Color ReportPrimaryColor = Colors.red;
+
+  static const Map<int, Color> MonthColorPalette = {
+    DateTime.january: Colors.red,
+    DateTime.february: Colors.pink,
+    DateTime.march: Colors.purple,
+    DateTime.april: Colors.indigo,
+    DateTime.may: Colors.blue,
+    DateTime.june: Colors.lightBlue,
+    DateTime.july: Colors.green,
+    DateTime.august: Colors.lime,
+    DateTime.september: Colors.yellow,
+    DateTime.october: Colors.orange,
+    DateTime.november: Colors.deepOrange,
+    DateTime.december: Colors.brown,
+  };
 }
